Add HTTP retries
During testing I've noticed the HTTP request fails on occasion. This unfortunately halts the entire script. To resolve this, I'm adding a HTTP retry mechanism for post requests. The idea is to allow for retries to occur when we get 5XX codes.

A small issue that may occur is that LGTM is not idempotent. That said, I think that for our use case we should be fine.