Add check for protoprojects
There isn't an easy way to see if a protoproject is eligible to be moved over to a custom list. To resolve this, we are checking to see if the project's language status is a success. This tells us that the protoproject succeeded in a build and is eligible to be moved.

Also, we are updating the self.id here. This is because of how LGTM treats protoprojects. A protoproject has a different type of id than a real project. However, as soon as a protoproject build succeeded, LGTM updates the protoproject's id. This is frustrating as it invalidates our cache record of the protoproject id. So to resolve this we update the protoproject id programmatically in-memory. It's nasty but it works.